homeduino with onewire multibus
This sketch is a extended version of homeduino.ino proposed by @ronaldsteen.

Now the DST command has 3 parameters (onewire_pin, sensor_number and precision). This adds possibilities to connect tens of sensors (may be used multiple bus). 

The bus is initialized each reading. It is not efficient but it is more easy (all in one command) and performance is good.
Can be improved of course. Do it, please.

Important: Pins for connecting busses can be any free pins. It's not recommended pin 2 and pin 4 because they are used as default receiver and transmitter. Be careful.


Regards!